+++ config/routes.rb
@@ -1,8 +1,11 @@
+# frozen_string_literal: true
+
 Rails.application.routes.draw do
+  root 'welcome#index'
 
   resources :videos do
     collection do
-      get "reorder/:video_page", action: 'reorder', as: :reorder
+      get 'reorder/:video_page', action: 'reorder', as: :reorder
       post :save_video_order
     end
   end
@@ -51,14 +54,14 @@ Rails.application.routes.draw do
     end
   end
 
-  get "invite/:invite_token" => "project_users#invite"
+  get 'invite/:invite_token' => 'project_users#invite'
 
   scope module: 'welcome' do
     get :dashboard
     get :index
     get :work
     get :drtv
-    get "images/videos/:video_id", action: :download_video_image, as: :download_video_image
+    get 'images/videos/:video_id', action: :download_video_image, as: :download_video_image
     get :contact
     post :submit_contact
     # get :news
@@ -72,30 +75,27 @@ Rails.application.routes.draw do
   end
 
   scope module: 'client_session' do
-    get  "client/login",  action: 'new',     as: :client_login
-    post "client/login",  action: 'create'
-    get  "client/logout", action: 'destroy', as: :client_logout
+    get  'client/login',  action: 'new',     as: :client_login
+    post 'client/login',  action: 'create'
+    get  'client/logout', action: 'destroy', as: :client_logout
   end
 
-  devise_for :users, path_names: { sign_up: 'join', sign_in: 'login', sign_out: 'logout' }, path: ""
+  devise_for :users, path_names: { sign_up: 'join', sign_in: 'login', sign_out: 'logout' }, path: ''
 
   resources :users
 
   scope module: 'client_project' do
-    get ":id", action: 'root', as: :client_project_root
-    get ":id/menu", action: 'menu', as: :client_project_menu
-    get ":id/agency_logo", action: 'agency_logo', as: :client_project_agency_logo
-    get ":id/client_logo", action: 'client_logo', as: :client_project_client_logo
+    get ':id', action: 'root', as: :client_project_root
+    get ':id/menu', action: 'menu', as: :client_project_menu
+    get ':id/agency_logo', action: 'agency_logo', as: :client_project_agency_logo
+    get ':id/client_logo', action: 'client_logo', as: :client_project_client_logo
 
-    get ":id/documents/download/:document_id", action: 'download_document', as: :client_project_download_document
-    get ":id/photos/download/:location_photo_id", action: 'download_location_photo', as: :client_project_download_location_photo
-    get ":id/:top_level/:category_id/photos/:location_id", action: :location_show, as: :client_project_location
-    get ":id/:top_level/:category_id/photos/:location_id/:location_photo_id", action: :location_photo, as: :client_project_location_photo
-    get ":id/:top_level/:category_id", action: 'category', as: :client_project_category
-    get ":id/:top_level/:category_id/:document_id", action: 'document', as: :client_project_category_document
-    get ":id/:category_id", action: 'category'
+    get ':id/documents/download/:document_id', action: 'download_document', as: :client_project_download_document
+    get ':id/photos/download/:location_photo_id', action: 'download_location_photo', as: :client_project_download_location_photo
+    get ':id/:top_level/:category_id/photos/:location_id', action: :location_show, as: :client_project_location
+    get ':id/:top_level/:category_id/photos/:location_id/:location_photo_id', action: :location_photo, as: :client_project_location_photo
+    get ':id/:top_level/:category_id', action: 'category', as: :client_project_category
+    get ':id/:top_level/:category_id/:document_id', action: 'document', as: :client_project_category_document
+    get ':id/:category_id', action: 'category'
   end
-
-  root 'welcome#index'
-
 end